+++ StabilityMatrix.Avalonia/ViewModels/Inference/ImageFolderCardViewModel.cs
@@ -383,12 +383,56 @@ public partial class ImageFolderCardViewModel : DisposableViewModelBase
     [RelayCommand]
     private async Task CopySeedToClipboard(LocalImageFile? item)
     {
-        if (item?.GenerationParameters is null)
+        if (item?.GenerationParameters is null || App.Clipboard is null)
         {
             return;
         }
 
         await App.Clipboard.SetTextAsync(item.GenerationParameters.Seed.ToString());
+    }
+
+    [RelayCommand]
+    private async Task CopyPromptToClipboard(LocalImageFile? item)
+    {
+        if (item?.GenerationParameters is null || App.Clipboard is null)
+        {
+            return;
+        }
+
+        await App.Clipboard.SetTextAsync(item.GenerationParameters.PositivePrompt);
+    }
+
+    [RelayCommand]
+    private async Task CopyNegativePromptToClipboard(LocalImageFile? item)
+    {
+        if (item?.GenerationParameters is null || App.Clipboard is null)
+        {
+            return;
+        }
+
+        await App.Clipboard.SetTextAsync(item.GenerationParameters.NegativePrompt);
+    }
+
+    [RelayCommand]
+    private async Task CopyModelNameToClipboard(LocalImageFile? item)
+    {
+        if (item?.GenerationParameters is null || App.Clipboard is null)
+        {
+            return;
+        }
+
+        await App.Clipboard.SetTextAsync(item.GenerationParameters.ModelName);
+    }
+
+    [RelayCommand]
+    private async Task CopyModelHashToClipboard(LocalImageFile? item)
+    {
+        if (item?.GenerationParameters is null || App.Clipboard is null)
+        {
+            return;
+        }
+
+        await App.Clipboard.SetTextAsync(item.GenerationParameters.ModelHash);
     }
 
     [RelayCommand]